updated download hooks
